requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
eth-account>=0.9.0
base58>=2.1.1
cryptography>=41.0.0
PyJWT>=2.8.0



